Day 21-22: changes to replay, rm and refactor   * Implemented rm features to pass cli tests   * Passed CLI tests   * Change replay logic to be less verbose   * Break module into multiple files   * Fix flakey tests
